Renamed to shorter names
